basic/CLockable.h: Remove PWX_*_OBJ() macros
These macros were once meant to enable using the safety macros with
pointers that could not be nullptr, so the usage of the macros fall
in line with the regular non _OBJ macros.

But this was a bad idea due to the following reasons:
* If a pointer can not be nullptr, just use it directly.
* Enforcing an unneccessary check is counter-productive.

They are gone for good now.

Signed-off-by: Sven Eden <sven.eden@prydeworx.com>